r/mutex_buffer: implemented buffering of continuations waiting for mutex
Implemented a class that allows to wrap mutex with a buffer. When
continuation chain wants to acquire a mutex it is put into the waiter
list, then after mutex is available each waiter is executed separately.
Mutex buffer allow all pending waiters to be executed when mutex is
available. This approach helps to reduce the consensus oplock
contention.

Signed-off-by: Michal Maslanka <michal@vectorized.io>