r/consensus: added decommission_node method
Added `consensus:decommission_node` method to mark nodes as
decommissioned in raft configuration.

Currently we use raft_0 configuration as a source of truth containing
information about the members of the cluster. In order to be consistent
and use the same mechanism for nodes decommissioning added an API to
raft::consensus allowing user to mark nodes as decommissioned.

NOTE: the same as brokers the decommissioned nodes doesn't have to be
stored as raft vnodes as they doesn't influence the raft protocol.

Signed-off-by: Michal Maslanka <michal@vectorized.io>