storage: do not steal head chunk on each write dispatch
Before this change when a write was dispatched from the _head chunk the
_head pointer to the chunk was cleared and then reinstated after the
write completed if the chunk were a partial chunk. If it were full then
it was returned to the chunk cache.

This worked fine when writes and flushes were never allowed to complete
in parallel. This patch changes the behavior to leave the chunk in place
but this leave the problem of dealing with multiple dispatched writes to
the same _head. This can happen for example:

   1. append(1 byte) <-- no dispatch occurrs
   2. flush          <-- dispatch occurs
   3. append(large)  <-- starts before (2) completes

Because the append in (3) is large the head chunk will fill up and be
dispatched as a new write, but it will be racing with a write to the
same chunk that were dispatched in (2).

To deal with this race we force writes to the same chunk to complete
in the order they were dispatched. This is generally not a problem as
most workloads dispatch large appends which consist primarily of full
chunks which do not require serialization.

Signed-off-by: Noah Watkins <noah@vectorized.io>