storage: support concurrent append and flushing
This patch adds support to the segment appender for allowing append() to
be called before the future from a previous flush() is completed. This
new functionality allows raft to pipeline appends while while waiting on
flushes to complete asynchronously.

The basic approach is that flushes register the offset that they want
flushed and as writes complete in the background they dispatch flushes
when necessary.

Signed-off-by: Noah Watkins <noah@vectorized.io>