r/recovery_stm: allowing multiple request to be sent from recovery_stm
Previously `recovery_stm` didn't leverage request pipelining. When
request was sent to follower it waited for the response before reading
next chunk of data and sending it to follower.

Changed `recovery_stm` logic to allow multiple requests to be sent to
the follower at the same time. This allows us to speed up followers
recovery. Additionally we interleave reading date with waiting for the
response. Follower is also able to coalesce FSYNCs when they are
required.

Signed-off-by: Michal Maslanka <michal@vectorized.io>