c/members_manager: wait for all changes before removing node
Wait for all pending partition configuration changes to finish before
performing the removal of decommissioned node. Previously it might have
happen that the node was removed before all the partitions were drained
and safely removed from it.

The problem manifested itself only when controller leadership changed
during node decommissioning. In this case the new raft0 leader might
already seen updated metadata i.e. the one which doesn't contain the
node that is being removed but configuration change may still be in
progress in `controller_backend`. Members backend is not aware of this
as current partition configuration doesn't contain the decommissioned
node already so it does not require monitoring. Waiting for all the
partition operations to finish guarantee that in progress changes will
be finished before performing the final removal step of decommissioned
node.

Signed-off-by: Michal Maslanka <michal@vectorized.io>