c/topics_table: store revision of partition replicas
When reverting partition operation we can not use the revision of
operation canceling a move. When configuration update that is being
canceled removes a replica from partition replica set when reverting
that operation replica is added back to partition configuration. In
order for the configurations to be equivalent, we must use exactly the
same revision which was used by replica that was removed.

Replicas revision information is stored in raft configurations, however
relaying on raft configuration itself is impossible as they may get
deleted or truncated (as a part of raft operations). Hence the need to
store replica nodes revisions in topic table.

This PR introduces a map that stores per partition replica revisions.
The state is built in memory from existing information applied to the
the topics table.

Signed-off-by: Michal Maslanka <michal@redpanda.com>