r/topic_table: pass replica revisions map in delta
Topic table tracks partition replica revisions. Revisions are updated as
soon as the command is applied to the topic table. Since controller
backend may need replica revision map from the time that the command was
applied we must store the snapshot of replica revisions map in
`topic_table_delta` this way controller backend does not have to query
topic table for the state. This would be incorrect as the state may not
be in align with the delta being processed by controller backend.

Signed-off-by: Michal Maslanka <michal@redpanda.com>