c/members_backend: do not try to move the same partitions when limited
When concurrent reallocations limit is reached the members backend
should continue iterating over the partitions list in the place where it
left of not to move the same set of partitions many times. In order to
achieve that we track the index of ntp that was previously processed.

Next reconciliation iteration will pick up where the previous one left
off. We can not store the topic map iterator as the map may change
between subsequent reconciliation loop runs. Additionally ntps are
stored in a hierarchical data structure. It would be wasteful to flatten
it out every time the balancer wants to iterate over list of ntps.

The last processed index is reset whenever reassignments loop finish
without reaching the max_reallocations limit. This allows multiple
reallocation attempts for the same update if needed.

Signed-off-by: Michal Maslanka <michal@redpanda.com>