r/consensus: introduced heartbeats v2
Introduced a new heartbeat format in Raft. Change the way how protocol
metadata are represented in heartbeats to reduce the overhead of
exchanging heartbeats between idle partitions.

The idea leading to optimisation in this PR is based on the observation
that when the partitions are quiescent information exchanged between the
leader and followers in heartbeat requests are exactly the same. This
allow us to reduce the amount of data sent from the leader and back
during those quiescent period of times. Lightweight heartbeat is reduced
to a group_id which indicates that the leader is alive and active.
Keeping leader heartbeats present, instead of not sending them at all
from the leader in quiescent state allow us to reduce changes in Raft
protocol implementation to minimum.

In the next step heartbeat_manager will be further optimised to stop
sending heartbeats to groups in quiescent state completely.

Implementation introduces a new endpoint in Raft service heartbeats_v2.
After all the nodes in the cluster are up to date the heartbeat_manger
automatically switch to use the new heartbeat version. Use of
lightweight heartbeat is additionally controller by a configuration flag
enable_lw_heartbeat.

Signed-off-by: Michal Maslanka <michal@redpanda.com>