cs/s3_client: url encode target path
S3 client calculates the signature over the request transformed into a
canonical for according to the AWS specification:

https://docs.aws.amazon.com/AmazonS3/latest/API/sig-v4-header-based-auth.html

It is critical for the canonical path to be exactly the same as the path
in the request otherwise the signature won't match.

This commit adds URL encoding to the S3 client in order to preserve all
already escaped characters from the original target. (The escaping of
the characters is required to meet the AWS S3 object key naming
limitations).

Previously the S3 client url encoding wasn't required and it all worked
correctly as Redpanda doesn't allow special characters in topic names.

Signed-off-by: Michał Maślanka <michal@redpanda.com>